Reimplement bootstrap, and add module magic to setup-env.sh
spack clone now contains the functionality of the old spack bootstrap.
spack bootstrap now installs needed packages for full functionality. A
list 'needed_specs' contains a list of specs for these packages and can
be expanded when more packages are recognized to be needed.

Add code to automatically enable the module function with setup-env.sh
if module isn't already available and environment-modules~X is
installed in spack.

When setting up a new spack installation the recommended procedure is
now:

. ${spack}/share/spack/setup-env.sh
spack bootstrap
. ${spack}/share/spack/setup-env.sh

module will then be available and ready to work.